Function sentry_anyhow::capture_anyhow

pub fn capture_anyhow(e: &Error) -> Uuid

Captures an anyhow::Error.

This will capture an anyhow error as a sentry event if a sentry::Client is initialised, otherwise it will be a no-op. The event is dispatched to the thread-local hub, with semantics as described in Hub::current.

See module level documentation for more information.